fix access problem

+++ ries-auth-service/src/main/java/com/revature/controller/AuthController.java
@@ -1,13 +1,16 @@
 package com.revature.controller;
 
+import com.revature.Force;
 import com.revature.model.Guest;
 import com.revature.model.RequisitionDTO;
 import com.revature.service.GuestService;
 import org.hibernate.exception.ConstraintViolationException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.oauth2.client.OAuth2RestTemplate;
+import org.springframework.security.oauth2.provider.OAuth2Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -16,13 +19,17 @@ import java.util.Random;
 @RestController
 public class AuthController {
     @Autowired
+    Force force;
+    @Autowired
     GuestService service;
 
     @Autowired
     OAuth2RestTemplate restTemplate;
 
     @RequestMapping(value = "/guests", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<List<Guest>> getAllGuests() {
+    public ResponseEntity getAllGuests(OAuth2Authentication auth) {
+        if(!force.getCurrentEmployee(auth).isRecruiter())
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Not a recruiter.");
         return ResponseEntity.ok(service.getAll());
     }
 
@@ -37,7 +44,9 @@ public class AuthController {
     }
 
     @RequestMapping(value = "/guest", method = RequestMethod.POST)
-    public ResponseEntity createGuest(@RequestBody GuestRequisitionHolder holder) {
+    public ResponseEntity createGuest(OAuth2Authentication auth, @RequestBody GuestRequisitionHolder holder) {
+        if(!force.getCurrentEmployee(auth).isRecruiter())
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Not a recruiter.");
         Guest guest = holder.guest;
         RequisitionDTO requisitionDTO = holder.requisition;
 
@@ -65,7 +74,9 @@ public class AuthController {
     }
 
     @RequestMapping(value = "/guest", method = RequestMethod.DELETE)
-    public ResponseEntity deleteGuest(@RequestBody int pin) {
+    public ResponseEntity deleteGuest(OAuth2Authentication auth, @RequestBody int pin) {
+        if(!force.getCurrentEmployee(auth).isRecruiter())
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Not a recruiter.");
         Guest guest = service.getByPin(pin);
         if(guest != null) {
             service.delete(guest);
@@ -78,5 +89,10 @@ public class AuthController {
     private class GuestRequisitionHolder {
         private Guest guest;
         private RequisitionDTO requisition;
+    }
+
+    @RequestMapping(value = "/test", method = RequestMethod.GET)
+    public OAuth2Authentication test(OAuth2Authentication auth) {
+        return auth;
     }
 }
